requirements.txt
# Core dependencies
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
jieba>=0.42.1
pymysql>=1.0.0
python-dotenv>=1.0.0

# Text processing
zhconv>=1.4.0
opencc-python-reimplemented>=0.1.7

# Performance and utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
click>=8.1.0
schedule>=1.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
